+++ src/main/kotlin/no/nav/helsearbeidsgiver/config/ApplicationConfig.kt
@@ -1,0 +1,185 @@
+package no.nav.helsearbeidsgiver.config
+
+import com.nimbusds.jose.util.DefaultResourceRetriever
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.auth.Authentication
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import no.nav.hag.utils.bakgrunnsjobb.PostgresBakgrunnsjobbRepository
+import no.nav.helsearbeidsgiver.Env.getProperty
+import no.nav.helsearbeidsgiver.Env.getPropertyOrNull
+import no.nav.helsearbeidsgiver.auth.AltinnAuthClient
+import no.nav.helsearbeidsgiver.auth.gyldigScope
+import no.nav.helsearbeidsgiver.auth.gyldigSystembrukerOgConsumer
+import no.nav.helsearbeidsgiver.bakgrunnsjobb.InnsendingProcessor
+import no.nav.helsearbeidsgiver.bakgrunnsjobb.LeaderElectedBakgrunnsjobbService
+import no.nav.helsearbeidsgiver.dialogporten.IDialogportenService
+import no.nav.helsearbeidsgiver.dialogporten.IngenDialogportenService
+import no.nav.helsearbeidsgiver.forespoersel.ForespoerselRepository
+import no.nav.helsearbeidsgiver.forespoersel.ForespoerselService
+import no.nav.helsearbeidsgiver.innsending.InnsendingRepository
+import no.nav.helsearbeidsgiver.innsending.InnsendingService
+import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingRepository
+import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingService
+import no.nav.helsearbeidsgiver.kafka.createKafkaConsumerConfig
+import no.nav.helsearbeidsgiver.kafka.createKafkaProducerConfig
+import no.nav.helsearbeidsgiver.kafka.forespoersel.ForespoerselTolker
+import no.nav.helsearbeidsgiver.kafka.innsending.InnsendingProducer
+import no.nav.helsearbeidsgiver.kafka.innsending.InnsendingSerializer
+import no.nav.helsearbeidsgiver.kafka.inntektsmelding.InntektsmeldingTolker
+import no.nav.helsearbeidsgiver.kafka.startKafkaConsumer
+import no.nav.helsearbeidsgiver.mottak.MottakRepository
+import no.nav.helsearbeidsgiver.pdp.IPdpService
+import no.nav.helsearbeidsgiver.pdp.IngenTilgangPdpService
+import no.nav.helsearbeidsgiver.pdp.LocalhostPdpService
+import no.nav.helsearbeidsgiver.pdp.PdpService
+import no.nav.helsearbeidsgiver.pdp.lagPdpClient
+import no.nav.security.token.support.core.configuration.ProxyAwareResourceRetriever.Companion.DEFAULT_HTTP_CONNECT_TIMEOUT
+import no.nav.security.token.support.core.configuration.ProxyAwareResourceRetriever.Companion.DEFAULT_HTTP_READ_TIMEOUT
+import no.nav.security.token.support.core.configuration.ProxyAwareResourceRetriever.Companion.DEFAULT_HTTP_SIZE_LIMIT
+import no.nav.security.token.support.v2.IssuerConfig
+import no.nav.security.token.support.v2.RequiredClaims
+import no.nav.security.token.support.v2.TokenSupportConfig
+import no.nav.security.token.support.v2.tokenValidationSupport
+import org.apache.kafka.clients.consumer.KafkaConsumer
+import org.apache.kafka.clients.producer.KafkaProducer
+import org.apache.kafka.common.serialization.StringSerializer
+import org.jetbrains.exposed.sql.Database
+
+data class Repositories(
+    val inntektsmeldingRepository: InntektsmeldingRepository,
+    val forespoerselRepository: ForespoerselRepository,
+    val mottakRepository: MottakRepository,
+    val innsendingRepository: InnsendingRepository,
+    val bakgrunnsjobbRepository: PostgresBakgrunnsjobbRepository,
+)
+
+data class Services(
+    val forespoerselService: ForespoerselService,
+    val inntektsmeldingService: InntektsmeldingService,
+    val innsendingService: InnsendingService,
+    val dialogportenService: IDialogportenService,
+)
+
+fun configureRepositories(db: Database): Repositories =
+    Repositories(
+        inntektsmeldingRepository = InntektsmeldingRepository(db),
+        forespoerselRepository = ForespoerselRepository(db),
+        mottakRepository = MottakRepository(db),
+        innsendingRepository = InnsendingRepository(db),
+        bakgrunnsjobbRepository = PostgresBakgrunnsjobbRepository(DbConfig.getDataSource()),
+    )
+
+fun configureServices(repositories: Repositories): Services {
+    val forespoerselService = ForespoerselService(repositories.forespoerselRepository)
+    val inntektsmeldingService = InntektsmeldingService(repositories.inntektsmeldingRepository)
+
+    val innsendingProducer =
+        InnsendingProducer(
+            KafkaProducer(
+                createKafkaProducerConfig(producerName = "api-innsending-producer"),
+                StringSerializer(),
+                InnsendingSerializer(),
+            ),
+        )
+
+    val bakgrunnsjobbService =
+        LeaderElectedBakgrunnsjobbService(
+            bakgrunnsjobbRepository = repositories.bakgrunnsjobbRepository,
+        ).apply {
+            registrer(InnsendingProcessor(repositories.innsendingRepository))
+            startAsync(true)
+        }
+
+    val innsendingService =
+        InnsendingService(
+            innsendingProducer = innsendingProducer,
+            innsendingRepository = repositories.innsendingRepository,
+            bakgrunnsjobbService = bakgrunnsjobbService,
+        )
+
+    // val dialogService = if (isDev()) DialogportenService(lagDialogportenClient(authClient)) else IngenDialogportenService()
+    val dialogportenService = IngenDialogportenService()
+
+    return Services(forespoerselService, inntektsmeldingService, innsendingService, dialogportenService)
+}
+
+fun Application.configureKafkaConsumers(
+    services: Services,
+    repositories: Repositories,
+) {
+    val kafkaEnabled = getProperty("kafkaConsumer.enabled").toBoolean()
+
+    if (kafkaEnabled) {
+        val inntektsmeldingKafkaConsumer = KafkaConsumer<String, String>(createKafkaConsumerConfig("im"))
+        launch(Dispatchers.Default) {
+            startKafkaConsumer(
+                getProperty("kafkaConsumer.inntektsmelding.topic"),
+                inntektsmeldingKafkaConsumer,
+                InntektsmeldingTolker(
+                    services.inntektsmeldingService,
+                    repositories.mottakRepository,
+                ),
+            )
+        }
+
+        val forespoerselKafkaConsumer = KafkaConsumer<String, String>(createKafkaConsumerConfig("fsp"))
+        launch(Dispatchers.Default) {
+            startKafkaConsumer(
+                getProperty("kafkaConsumer.forespoersel.topic"),
+                forespoerselKafkaConsumer,
+                ForespoerselTolker(
+                    repositories.forespoerselRepository,
+                    repositories.mottakRepository,
+                    services.dialogportenService,
+                ),
+            )
+        }
+    }
+}
+
+fun Application.configureAuth() {
+    val authClient = AltinnAuthClient()
+    val pdpService = configurePdpService(authClient)
+
+    install(Authentication) {
+        tokenValidationSupport(
+            name = "systembruker-config",
+            config =
+                TokenSupportConfig(
+                    IssuerConfig(
+                        name = "maskinporten",
+                        discoveryUrl = getProperty("maskinporten.wellknownUrl"),
+                        acceptedAudience = listOf(getProperty("maskinporten.eksponert_scopes")),
+                        optionalClaims = listOf("aud", "sub"),
+                    ),
+                ),
+            requiredClaims =
+                RequiredClaims(
+                    issuer = "maskinporten",
+                    claimMap = arrayOf("authorization_details", "consumer", "scope"),
+                ),
+            additionalValidation = {
+                it.gyldigScope() && it.gyldigSystembrukerOgConsumer(pdpService)
+            },
+            resourceRetriever =
+                DefaultResourceRetriever(
+                    DEFAULT_HTTP_CONNECT_TIMEOUT,
+                    DEFAULT_HTTP_READ_TIMEOUT,
+                    DEFAULT_HTTP_SIZE_LIMIT,
+                ),
+        )
+    }
+}
+
+fun configurePdpService(authClient: AltinnAuthClient): IPdpService =
+    when {
+        isDev() -> PdpService(lagPdpClient(authClient))
+        isLocal() -> LocalhostPdpService()
+        else -> IngenTilgangPdpService()
+    }
+
+private fun isDev(): Boolean = "dev-gcp".equals(getPropertyOrNull("application.env"), true)
+
+private fun isLocal(): Boolean = "local".equals(getPropertyOrNull("application.env"), true)
