+++ src/main/kotlin/no/nav/helsearbeidsgiver/plugins/Routing.kt
@@ -4,24 +4,18 @@ import io.ktor.server.application.Application
 import io.ktor.server.auth.authenticate
 import io.ktor.server.plugins.swagger.swaggerUI
 import io.ktor.server.routing.routing
-import no.nav.helsearbeidsgiver.forespoersel.ForespoerselService
+import no.nav.helsearbeidsgiver.config.Services
 import no.nav.helsearbeidsgiver.forespoersel.forespoerselV1
-import no.nav.helsearbeidsgiver.innsending.InnsendingService
 import no.nav.helsearbeidsgiver.innsending.innsendingV1
-import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingService
 import no.nav.helsearbeidsgiver.inntektsmelding.inntektsmeldingV1
 
-fun Application.configureRouting(
-    forespoerselService: ForespoerselService,
-    inntektsmeldingService: InntektsmeldingService,
-    innsendingService: InnsendingService,
-) {
+fun Application.configureRouting(services: Services) {
     routing {
         swaggerUI(path = "swagger", swaggerFile = "openapi/documentation.yaml")
         authenticate("systembruker-config") {
-            inntektsmeldingV1(inntektsmeldingService)
-            forespoerselV1(forespoerselService)
-            innsendingV1(innsendingService)
+            inntektsmeldingV1(inntektsmeldingService = services.inntektsmeldingService)
+            forespoerselV1(forespoerselService = services.forespoerselService)
+            innsendingV1(innsendingService = services.innsendingService)
         }
     }
 }
