+++ src/test/kotlin/no/nav/helsearbeidsgiver/kafka/inntektsmelding/MeldingTolkerTest.kt
@@ -3,7 +3,7 @@ package no.nav.helsearbeidsgiver.kafka.inntektsmelding
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
-import no.nav.helsearbeidsgiver.db.Database
+import no.nav.helsearbeidsgiver.config.DbConfig
 import no.nav.helsearbeidsgiver.dialogporten.IDialogportenService
 import no.nav.helsearbeidsgiver.forespoersel.ForespoerselRepository
 import no.nav.helsearbeidsgiver.inntektsmelding.InntektsmeldingRepository
@@ -22,7 +22,7 @@ import java.util.UUID
 
 @ExtendWith(TransactionalExtension::class)
 class MeldingTolkerTest {
-    val db = Database.init()
+    val db = DbConfig.init()
     val inntektsmeldingRepository = InntektsmeldingRepository(db)
     val forespoerselRepository = ForespoerselRepository(db)
     val inntektsmeldingService = InntektsmeldingService(inntektsmeldingRepository)
