+++ src/test/resources/application.conf
@@ -11,7 +11,7 @@ maskinporten {
   eksponert_scopes = ${?EKSPONERT_MASKINPORTEN_SCOPES}
 }
 database {
-  embedded = true
+  embedded = false
 }
 application {
   env = "local"
